requirements.txt
alembic==1.17.2
annotated-types==0.7.0
bcrypt==4.3.0
fastapi==0.116.1
greenlet==3.3.0
langchain==0.3.27
langchain_google_genai==2.1.8
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.1.1
SQLAlchemy==2.0.45
typing-inspection==0.4.2
typing_extensions==4.15.0
